feat(dsm-announcements): add member validation

+++ src/services/dsm/announcements.services.js
@@ -7,7 +7,8 @@ const selectOnlyValidAnnouncementFields = {
     author: true,
     content: true,
     createdAt: true,
-    projectId: true
+    projectId: true,
+    memberId: true,
   }
 };
 
@@ -15,8 +16,11 @@ const selectOnlyValidAnnouncementFields = {
   * Service to list all the announcements
   * @returns {Object} - List of announcements
 */
-const getAnnouncements = async () => {
+const getAnnouncements = async (projectId) => {
   const announcements = await dashboardPrisma.Announcement.findMany({
+    where : {
+      projectId
+    },
     orderBy: {
       createdAt: 'desc',
     },
@@ -34,7 +38,7 @@ const getAnnouncements = async () => {
 const getAnnouncementByID = async (announcementId) => {
   const announcement = await dashboardPrisma.Announcement.findUnique({
     where: {
-      announcementId
+      announcementId,
     },
     ...selectOnlyValidAnnouncementFields
   });
@@ -50,12 +54,13 @@ const getAnnouncementByID = async (announcementId) => {
   * @param {String} content - content of the announcement
   * @returns {Object} - Announcement object
 */
-const createAnnouncement = async (author, content, projectId) => {
+const createAnnouncement = async (author, memberId, content, projectId) => {
   const announcement = await dashboardPrisma.Announcement.create({
     data: {
-      author,
       content,
-      projectId
+      projectId,
+      memberId,
+      author
     },
     ...selectOnlyValidAnnouncementFields
   });
@@ -69,16 +74,23 @@ const createAnnouncement = async (author, content, projectId) => {
   * @returns {Object} - Updated announcement object
   * @throws {HttpError} - Throws an error if announcement not found
 */
-const editAnnouncement = async (announcementId, content) => {
+const editAnnouncement = async (announcementId, content, memberId) => {
   const availableAnnouncement = await dashboardPrisma.Announcement.findUnique({
     where: {
       announcementId
     },
     ...selectOnlyValidAnnouncementFields
   });
+
   if (!availableAnnouncement) {
     throw new HttpError(404, 'Announcement not found');
   }
+
+  // check if the user is the author of the announcement
+  if (availableAnnouncement.memberId !== memberId) {
+    throw new HttpError(403, 'You are not authorised to edit this announcement');
+  }
+
   //TODO: Can be optimised further
   const announcement = await dashboardPrisma.Announcement.update({
     where: {
@@ -98,16 +110,22 @@ const editAnnouncement = async (announcementId, content) => {
   * @returns {void} - No return value
   * @throws {HttpError} - Throws an error if announcement not found
 */
-const deleteAnnouncement = async (announcementId) => {
+const deleteAnnouncement = async (announcementId, memberId) => {
   const announcement = await dashboardPrisma.Announcement.findUnique({
     where: {
       announcementId
     },
     ...selectOnlyValidAnnouncementFields
   });
+
   if (!announcement) {
     throw new HttpError(404, 'Announcement not found');
   }
+
+  if (announcement.memberId !== memberId) {
+    throw new HttpError(403, 'You are not authorised to delete this announcement');
+  }
+
   // TODO: Can be optimised further
   await dashboardPrisma.Announcement.delete({
     where: {
